📦 🐛 add  fix in  broadcast list

+++ src/app/main/pages/quality/configurations/configurations.component.ts
@@ -1,6 +1,12 @@
 import { BreakpointObserver, Breakpoints } from '@angular/cdk/layout';
 import { Component, OnInit } from '@angular/core';
-import { FormBuilder, FormControl, FormGroup, Validators, FormArray } from '@angular/forms';
+import {
+  FormBuilder,
+  FormControl,
+  FormGroup,
+  Validators,
+  FormArray,
+} from '@angular/forms';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { BehaviorSubject, Subscription, Observable } from 'rxjs';
 import { switchMap, take, tap } from 'rxjs/operators';
@@ -8,39 +14,36 @@ import { AuthService } from 'src/app/auth/services/auth.service';
 import { User } from 'src/app/main/models/user-model';
 import { QualityService } from 'src/app/main/services/quality.service';
 import { MyErrorStateMatcher } from '../../evaluations/evaluations-settings/evaluations-settings.component';
-import { QualityListSpecialist, QualityListResponsibleArea, QualityBroadcastList } from '../../../models/quality.model';
+import {
+  QualityListSpecialist,
+  QualityListResponsibleArea,
+  QualityBroadcastList,
+} from '../../../models/quality.model';
 
 @Component({
   selector: 'app-configurations',
   templateUrl: './configurations.component.html',
-  styleUrls: ['./configurations.component.scss']
+  styleUrls: ['./configurations.component.scss'],
 })
 export class ConfigurationsComponent implements OnInit {
-
   loading = new BehaviorSubject<boolean>(false);
   loading$ = this.loading.asObservable();
   panelOpenState = false;
-  listBahiasForm: FormGroup;
 
-  diffusionForm = new FormArray([]);
-  listDiffusionControl = new FormControl(null, [Validators.required]);
-  listDiffusionArray = [];
+  broadcastFormArray = new FormArray([]);
   broadcastListArray: QualityBroadcastList[] = [];
 
   listSpecialistControl = new FormControl(null, [Validators.required]);
   listSpecialistArray = [];
 
-  listResponsibleAreasControl  = new FormControl(null, [Validators.required]);
+  listResponsibleAreasControl = new FormControl(null, [Validators.required]);
   listResponsibleAreasArray = [];
-
-
 
   matcher = new MyErrorStateMatcher();
   listProblemTypeArray = [];
   listNameBahiaArray = [];
 
-  broadcast$: Observable<QualityBroadcastList[] >;
-
+  broadcast$: Observable<QualityBroadcastList[]>;
 
   private subscription = new Subscription();
   user: User;
@@ -54,20 +57,21 @@ export class ConfigurationsComponent implements OnInit {
     private snackbar: MatSnackBar,
     private qualityService: QualityService,
     private breakpoint: BreakpointObserver
-
-  ) { }
+  ) {}
 
   ngOnInit(): void {
-    this.subscription.add(this.breakpoint.observe([Breakpoints.HandsetPortrait])
-      .subscribe(res => {
-        if (res.matches) {
-          this.isMobile = true;
-          this.setHandsetContainer();
-        } else {
-          this.isMobile = false;
-          this.setDesktopContainer();
-        }
-      })
+    this.subscription.add(
+      this.breakpoint
+        .observe([Breakpoints.HandsetPortrait])
+        .subscribe((res) => {
+          if (res.matches) {
+            this.isMobile = true;
+            this.setHandsetContainer();
+          } else {
+            this.isMobile = false;
+            this.setDesktopContainer();
+          }
+        })
     );
 
     this.broadcast$ = this.qualityService.getAllBroadcastList().pipe(
@@ -75,14 +79,20 @@ export class ConfigurationsComponent implements OnInit {
         if (res) {
           this.broadcastListArray = res;
         }
+        res.map((el) => {
+          this.broadcastFormArray.push(
+            new FormControl(null, Validators.required)
+          );
+        });
       })
-    )
-
+    );
 
     this.loading.next(true);
-    this.subscription.add(this.auth.user$.subscribe(user => {
-      this.user = user;
-    }));
+    this.subscription.add(
+      this.auth.user$.subscribe((user) => {
+        this.user = user;
+      })
+    );
 
     this.subscription.add(
       this.qualityService.getAllQualityListSpecialist().subscribe((resp) => {
@@ -95,44 +105,47 @@ export class ConfigurationsComponent implements OnInit {
     );
 
     this.subscription.add(
-      this.qualityService.getAllQualityListResponsibleAreas()
-      .subscribe((resp) => {
-        if (resp) {
-          this.listResponsibleAreasArray = resp;
-        } else {
-          this.listResponsibleAreasArray = [];
-        }
-      })
+      this.qualityService
+        .getAllQualityListResponsibleAreas()
+        .subscribe((resp) => {
+          if (resp) {
+            this.listResponsibleAreasArray = resp;
+          } else {
+            this.listResponsibleAreasArray = [];
+          }
+        })
     );
 
     this.loading.next(false);
+
   }
-  
+
   ngOnDestroy(): void {
     this.subscription.unsubscribe();
   }
 
-  addDiffusion(): void{
+  addDiffusion(): void {
     const lenghtArray = this.broadcastListArray.length;
     const nameBroadcast = `difusion ${lenghtArray + 1}`;
 
-    const arrayList: QualityBroadcastList =
-      {
-        id: null,
-        name: nameBroadcast,
-        emailList: null,
-        createdAt: null,
-        createdBy: null
-      };
+    const arrayList: QualityBroadcastList = {
+      id: null,
+      name: nameBroadcast,
+      emailList: null,
+      createdAt: null,
+      createdBy: null,
+    };
+    this.broadcastFormArray.push(new FormControl(null, Validators.required));
 
     this.broadcastListArray.push(arrayList);
-
   }
 
   addListDiffusion(broadcast: QualityBroadcastList, index: number): void {
+  
     try {
-      const name = broadcast.name;
-      const newBroadcast = this.listDiffusionControl.value.trim().toLowerCase();
+      let name = broadcast.name;
+      let newBroadcast = this.broadcastFormArray.controls[index].value.trim().toLowerCase();
+
       if (broadcast.id === null) {
         const resp = this.qualityService.addNewBrodcastList(newBroadcast, name, this.user);
         //this.loading.next(true);
@@ -145,8 +158,7 @@ export class ConfigurationsComponent implements OnInit {
                   this.snackbar.open('✅ se guardo correctamente!', 'Aceptar', {
                     duration: 6000
                   });
-                  this.listDiffusionControl.reset();
-
+                  this.broadcastFormArray.removeAt(index);
                 })
                 .catch(err => {
                 // this.loading.next(false);
@@ -172,12 +184,12 @@ export class ConfigurationsComponent implements OnInit {
                   this.snackbar.open('✅ se guardo correctamente!', 'Aceptar', {
                     duration: 6000
                   });
-                  this.listDiffusionControl.reset();
+                  this.broadcastFormArray.removeAt(index);
 
                 })
                 .catch(err => {
                 // this.loading.next(false);
-                  this.snackbar.open('🚨 Hubo un error al crear!', 'Aceptar', {
+                  this.snackbar.open('🚨 Hubo un error al actualizar  !', 'Aceptar', {
                     duration: 6000
                   });
                 });
@@ -193,45 +205,60 @@ export class ConfigurationsComponent implements OnInit {
 
   }
 
-  async deleteListBroadcast(broadcast: QualityBroadcastList): Promise<void> {
+  async deleteListBroadcast(
+    broadcast: QualityBroadcastList,
+    index: number
+  ): Promise<void> {
+    console.log('broadcast:', broadcast);
+    console.log('index:', index);
+    if (broadcast.id != null) {
       this.loading.next(true);
       await this.qualityService.deleteListBroadcast(broadcast.id);
       this.snackbar.open('✅ Elemento borrado correctamente', 'Aceptar', {
-        duration: 6000
+        duration: 6000,
       });
       this.loading.next(false);
+    } else {
+      this.broadcastListArray.splice(index, 1);
+    }
   }
 
-
-  updateBrocastListEmail(broadcast: QualityBroadcastList, broadcastList: string): void {
+  updateBrocastListEmail(
+    broadcast: QualityBroadcastList,
+    broadcastList: string
+  ): void {
     try {
-      const resp = this.qualityService.updateBrodcastEmailList(broadcast.id, broadcastList );
+      const resp = this.qualityService.updateBrodcastEmailList(
+        broadcast.id,
+        broadcastList
+      );
       //this.loading.next(true);
-      this.subscription.add(resp.subscribe(
-        batch => {
+      this.subscription.add(
+        resp.subscribe((batch) => {
           if (batch) {
-            batch.commit()
+            batch
+              .commit()
               .then(() => {
-               // this.loading.next(false);
+                // this.loading.next(false);
                 this.snackbar.open('✅ se elimino correctamente!', 'Aceptar', {
-                  duration: 6000
+                  duration: 6000,
                 });
               })
-              .catch(err => {
-               // this.loading.next(false);
+              .catch((err) => {
+                // this.loading.next(false);
                 this.snackbar.open('🚨 Hubo un error al crear!', 'Aceptar', {
-                  duration: 6000
+                  duration: 6000,
                 });
               });
           }
-        }
-      ));
+        })
+      );
     } catch (error) {
       console.log(error);
       this.loading.next(false);
     }
   }
- 
+
   addListSpecialist(): void {
     if (this.listSpecialistControl.valid) {
       const objAux: QualityListSpecialist = {
@@ -240,7 +267,8 @@ export class ConfigurationsComponent implements OnInit {
         createdAt: null,
         createdBy: null,
       };
-      const valueIsEquals = (currentValue) => currentValue.specialist !== objAux.specialist;
+      const valueIsEquals = (currentValue) =>
+        currentValue.specialist !== objAux.specialist;
       if (this.listSpecialistArray.every(valueIsEquals)) {
         this.listSpecialistArray.push(objAux);
       }
@@ -250,56 +278,64 @@ export class ConfigurationsComponent implements OnInit {
   async deleteListSpecialist(index: number): Promise<void> {
     if (this.listSpecialistArray[index].id) {
       this.loading.next(true);
-      await this.qualityService.deleteQualityListSpecialist(this.listSpecialistArray[index].id);
+      await this.qualityService.deleteQualityListSpecialist(
+        this.listSpecialistArray[index].id
+      );
       this.snackbar.open('✅ Elemento borrado correctamente', 'Aceptar', {
-        duration: 6000
+        duration: 6000,
       });
       this.loading.next(false);
     }
 
     this.listSpecialistArray.splice(index, 1);
     this.loading.next(false);
-
   }
 
   saveListSpecialist(): void {
-     try {
-       const resp = this.qualityService.addQualityListSpecialist(this.listSpecialistArray, this.user);
-       //this.loading.next(true);
-       this.subscription.add(resp.subscribe(
-         batch => {
-           if (batch) {
-             batch.commit()
-               .then(() => {
+    try {
+      const resp = this.qualityService.addQualityListSpecialist(
+        this.listSpecialistArray,
+        this.user
+      );
+      //this.loading.next(true);
+      this.subscription.add(
+        resp.subscribe((batch) => {
+          if (batch) {
+            batch
+              .commit()
+              .then(() => {
                 // this.loading.next(false);
-                 this.snackbar.open('✅ se guardo correctamente!', 'Aceptar', {
-                   duration: 6000
-                 });
-               })
-               .catch(err => {
+                this.snackbar.open('✅ se guardo correctamente!', 'Aceptar', {
+                  duration: 6000,
+                });
+              })
+              .catch((err) => {
                 // this.loading.next(false);
-                 this.snackbar.open('🚨 Hubo un error al crear!', 'Aceptar', {
-                   duration: 6000
-                 });
-               });
-           }
-         }
-       ));
-     } catch (error) {
-       console.log(error);
-       this.loading.next(false);
-     }
-   }
+                this.snackbar.open('🚨 Hubo un error al crear!', 'Aceptar', {
+                  duration: 6000,
+                });
+              });
+          }
+        })
+      );
+    } catch (error) {
+      console.log(error);
+      this.loading.next(false);
+    }
+  }
 
   addListResponsibleArea(): void {
     if (this.listResponsibleAreasControl.valid) {
       const objAux: QualityListResponsibleArea = {
         id: null,
-        responsable: this.listResponsibleAreasControl.value.trim().toLowerCase(),
+        responsable: this.listResponsibleAreasControl.value
+          .trim()
+          .toLowerCase(),
         createdAt: null,
         createdBy: null,
       };
-      const valueIsEquals = (currentValue) => currentValue.responsable !== objAux.responsable;
+      const valueIsEquals = (currentValue) =>
+        currentValue.responsable !== objAux.responsable;
       if (this.listResponsibleAreasArray.every(valueIsEquals)) {
         this.listResponsibleAreasArray.push(objAux);
       }
@@ -309,58 +345,61 @@ export class ConfigurationsComponent implements OnInit {
   async deleteListResponsibleArea(index: number): Promise<void> {
     if (this.listResponsibleAreasArray[index].id) {
       this.loading.next(true);
-      await this.qualityService.deleteQualityListResponsibleAreas(this.listResponsibleAreasArray[index].id);
+      await this.qualityService.deleteQualityListResponsibleAreas(
+        this.listResponsibleAreasArray[index].id
+      );
       this.snackbar.open('✅ Elemento borrado correctamente', 'Aceptar', {
-        duration: 6000
+        duration: 6000,
       });
       this.loading.next(false);
     }
 
     this.listResponsibleAreasArray.splice(index, 1);
     this.loading.next(false);
-
   }
 
   saveListResponsibleArea(): void {
-     try {
-       const resp = this.qualityService.addQualityListResponsibleAreas(this.listResponsibleAreasArray, this.user);
-       //this.loading.next(true);
-       this.subscription.add(resp.subscribe(
-         batch => {
-           if (batch) {
-             batch.commit()
-               .then(() => {
-                 //this.loading.next(false);
-                 this.snackbar.open('✅ se guardo correctamente!', 'Aceptar', {
-                   duration: 6000
-                 });
-               })
-               .catch(err => {
-                 this.loading.next(false);
-                 this.snackbar.open('🚨 Hubo un error al crear!', 'Aceptar', {
-                   duration: 6000
-                 });
-               });
-           }
-         }
-       ));
-     } catch (error) {
-       console.log(error);
-       this.loading.next(false);
-     }
-   }
-
+    try {
+      const resp = this.qualityService.addQualityListResponsibleAreas(
+        this.listResponsibleAreasArray,
+        this.user
+      );
+      //this.loading.next(true);
+      this.subscription.add(
+        resp.subscribe((batch) => {
+          if (batch) {
+            batch
+              .commit()
+              .then(() => {
+                //this.loading.next(false);
+                this.snackbar.open('✅ se guardo correctamente!', 'Aceptar', {
+                  duration: 6000,
+                });
+              })
+              .catch((err) => {
+                this.loading.next(false);
+                this.snackbar.open('🚨 Hubo un error al crear!', 'Aceptar', {
+                  duration: 6000,
+                });
+              });
+          }
+        })
+      );
+    } catch (error) {
+      console.log(error);
+      this.loading.next(false);
+    }
+  }
 
   setHandsetContainer(): void {
     this.containerStyle = {
-      'margin': '30px 24px 30px 24px'
-    }
+      margin: '30px 24px 30px 24px',
+    };
   }
 
   setDesktopContainer(): void {
     this.containerStyle = {
-      'margin': '30px 80px 30px 80px',
-    }
+      margin: '30px 80px 30px 80px',
+    };
   }
-
 }
